move work to lib
