--- app/src/temp/index.ts
@@ -1,8 +1,0 @@
-export { ActionModal, type ActionModalProps } from './ActionModal';
-export {
-  ActionModalProvider,
-  type ActionModalContextValue,
-  type AlertOptions,
-  type ConfirmOptions,
-} from './ActionModalProvider';
-export { useActionModal } from './useActionModal';
