+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.4-test.10",
+  "version": "1.7.4-test.11",
   "publishConfig": {
     "access": "public"
   },
